feat(tasks): Implement tasks folder structure and break up current grunt tasks fixes #84
