+++ config/index.js
@@ -1,0 +1,3 @@
+var path = require('path');
+
+module.exports = require(path.resolve(path.join(__dirname, '..', 'lib', 'config')));
