+++ lib/utils/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  helpers: require('./helpers')
+};
