+++ package.json
@@ -75,7 +75,10 @@
     "ng-annotate": "~0.9.11",
     "phantomjs": "~1.9.2",
     "protractor": "~0.24.2",
-    "time-grunt": "~0.2.1"
+    "time-grunt": "~0.2.1",
+    "deepmerge": "~0.2.10",
+    "debug": "~2.2.0",
+    "nconf": "~0.7.1"
   },
   "scripts": {
     "setup": "node scripts/setup",
