+++ tasks/grunt/concurrent.js
@@ -1,0 +1,11 @@
+module.exports = {
+  watch: {
+    tasks:          [
+      'watch:livereload',
+      'watch:less'
+    ],
+    options:        {
+      logConcurrentOutput: true
+    }
+  }
+};
