+++ tasks/grunt/makeReport.js
@@ -1,0 +1,8 @@
+module.exports = {
+  src:              '<%= appConfig.test.e2e.instrumented.path %>*.json',
+  options: {
+    type:           'html',
+    dir:            '<%= appConfig.test.e2e.coverage.path %>reports',
+    print:          'detail'
+  }
+};
