+++ tasks/grunt/ngAnnotate.js
@@ -1,0 +1,9 @@
+module.exports = {
+  options: {
+    singleQuotes:   true,
+    regexp:         '^(ng\n?[\\ ]+(.*)|(module.*))$'
+  },
+  dist: {
+    files:          [ { add: true, src: 'dist/modules/**/*.js' } ]
+  }
+};
