+++ tasks/grunt/rev.js
@@ -1,0 +1,12 @@
+module.exports = {
+  dist: {
+    files: {
+      src:          [
+        '<%= appConfig.dist.path %>/scripts/{,*/}*.js',
+        '<%= appConfig.dist.path %>/styles/{,*/}*.css',
+        '<%= appConfig.dist.path %>/images/{,*/}*.{png,jpg,jpeg,gif,webp,svg}',
+        '<%= appConfig.dist.path %>/styles/fonts/*'
+      ]
+    }
+  }
+};
