+++ tasks/grunt/useminPrepare.js
@@ -1,0 +1,7 @@
+module.exports = {
+  html:             '<%= appConfig.dev.path %>/index.html',
+  options: {
+    root:           '<%= appConfig.dev.path %>',
+    dest:           '<%= appConfig.dist.path %>'
+  }
+};
